:truck: Moved `ValueIterator` class to a separate file
