+++ src/ECS/utilities/ValueIterator.hpp
@@ -1,0 +1,25 @@
+/*
+** EPITECH PROJECT, 2024
+** BsRtype
+** File description:
+** ValueIterator.hpp file
+*/
+
+#include <utility>
+
+/**
+ * @brief Helper class for iterating over the values of a map
+ *
+ * @tparam The iterator type, for example: `std::map<int, int>::iterator`
+ *
+ * Inspired by https://stackoverflow.com/a/23608376
+ */
+template<typename It>
+struct ValueIterator : public It {
+    ValueIterator(It src) : It(std::move(src)) {}
+
+    auto &operator*() const 
+    {
+        return It::operator*().second;
+    }
+};
